ispiti/racunarstvo/20220129a/myimgur/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-bootstrap-datepicker-plus==3.0.5
django-widget-tweaks==1.4.8
pytz==2020.4
sqlparse==0.4.1
